通信协议修正，发送消息接口搞定，修改主控代码MyInfo的ID到8位，未验证

+++ Codes/主控/Strategy/strategy.h
@@ -31,8 +31,7 @@ typedef struct
 {
 	double Rect_Axis[2];
 	double Polar_Axis[2];
-	uint16_t MyID;
-	
+	u8 MyID;
 }MyInfo;
 /* Exported constants --------------------------------------------------------*/
 /* Exported macro ------------------------------------------------------------*/
